Update: split jsonHandler.ts to different files
