+++ src/commands/capacity.ts
@@ -1,6 +1,6 @@
 import { ChannelType, Client, CommandInteraction, ApplicationCommandOptionType } from 'discord.js'
 import { SlashCommand } from '../utils/commands.js'
-import { openConfig } from '../utils/jsonHandler.js'
+import { openConfig } from '../utils/index.js'
 
 export const Capacity: SlashCommand = {
     name: 'modify-capacity',
