+++ src/commands/messageStream.ts
@@ -1,6 +1,6 @@
 import { ApplicationCommandOptionType, ChannelType, Client, CommandInteraction } from 'discord.js'
 import { SlashCommand } from '../utils/commands.js'
-import { openConfig } from '../utils/jsonHandler.js'
+import { openConfig } from '../utils/index.js'
 
 export const MessageStream: SlashCommand = {
     name: 'message-stream',
