+++ src/commands/threadPrivateCreate.ts
@@ -1,6 +1,6 @@
 import { ChannelType, Client, CommandInteraction, TextChannel } from 'discord.js'
 import { SlashCommand } from '../utils/commands.js'
-import { openThreadInfo } from '../utils/jsonHandler.js'
+import { openThreadInfo } from '../utils/chatHistoryHandler.js'
 
 export const PrivateThreadCreate: SlashCommand = {
     name: 'private-thread',
