+++ src/utils/index.ts
@@ -5,3 +5,6 @@ export * from './messageEmbed.js'
 export * from './messageNormal.js'
 export * from './commands.js'
 export * from './streamHandler.js'
+export * from './fileTypes.js'
+export * from './chatHistoryHandler.js'
+export * from './configHandler.js'
